Add tsconfig, refactor src/
