+++ packages/council-delegates/src/types.ts
@@ -1,0 +1,18 @@
+export interface Delegate {
+  address: string;
+  commonwealthPostedFromAddress?: string;
+  commonwealthCommentId?: number;
+  commonwealthName?: string | null;
+  name?: string;
+  description?: string;
+
+  /**
+   * Twitter handle w/out the @ symbol, eg: "CharlieStLouis" (not "@CharlieStLouis")
+   */
+  twitterHandle?: string;
+
+  /**
+   * Timestamp of when the post was made
+   */
+  createdAt: string;
+}
